More work on database parsing
From now on, there will be 3 independent sets of ids: one for customers,
one for media objects, and one for purchase ids. I changed all string
type ids to ints. No more identifying chars in front of the id. ID 0 is
reserved for all 3 databases, so feel free to use it as a sort of back
or abort option when prompting for an id.